requirements.txt
anyio==3.5.0
click==8.1.7
fastapi~=0.109.0
freetype-py==2.1.0.post1
glfw==1.11.2
h11==0.12.0
hsluv==5.0.3
idna==3.4
kiwisolver==1.4.4
p5==0.8.2
pip==23.3.1
pybind11==2.10.4
pydantic==1.10.12
PyOpenGL==3.1.6
PyOpenGL-accelerate==3.1.6
python-multipart==0.0.6
requests==2.31.0
setuptools==68.2.2
skia-python==87.5
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.9.0
uvicorn==0.20.0
vispy==0.10.0
wheel==0.41.2
python-dotenv
removebg==0.4
